LayoutTests/http/tests/security/local-iFrame-from-remote-expected.txt

CONSOLE MESSAGE: line 0: Not allowed to load local resource: file:///tmp/LayoutTests/http/tests/security/resources/localPage.html
This test is to see if a remote file can include a local page in an iFrame.

Test Passed.
